pyproject.toml
[tool.poetry]
name = "my-health-checker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"
pync = "^2.0.3"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
